Removed unnecessary casting (5)

+++ SF3Lib/Models/Structs/MPD/Model/AttrStruct.cs
@@ -27,7 +27,7 @@ namespace SF3.Models.Structs.MPD.Model {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_planeAddr), displayOrder: 0, displayFormat: "X2")]
         public byte Plane {
-            get => (byte) Data.GetUInt8(_planeAddr);
+            get => Data.GetUInt8(_planeAddr);
             set => Data.SetUInt8(_planeAddr, value);
         }
 
@@ -41,7 +41,7 @@ namespace SF3.Models.Structs.MPD.Model {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_sortAndOptionsAddr), displayOrder: 0.02f, displayFormat: "X2")]
         public byte SortAndOptions {
-            get => (byte) Data.GetUInt8(_sortAndOptionsAddr);
+            get => Data.GetUInt8(_sortAndOptionsAddr);
             set => Data.SetUInt8(_sortAndOptionsAddr, value);
         }
 
@@ -69,14 +69,14 @@ namespace SF3.Models.Structs.MPD.Model {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_textureNoAddr), displayOrder: 2, displayFormat: "X4")]
         public ushort TextureNo {
-            get => (ushort) Data.GetUInt16(_textureNoAddr);
+            get => Data.GetUInt16(_textureNoAddr);
             set => Data.SetUInt16(_textureNoAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_modeAddr), displayOrder: 3, displayFormat: "X4")]
         public ushort Mode {
-            get => (ushort) Data.GetUInt16(_modeAddr);
+            get => Data.GetUInt16(_modeAddr);
             set => Data.SetUInt16(_modeAddr, value);
         }
 
@@ -137,7 +137,7 @@ namespace SF3.Models.Structs.MPD.Model {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_colorNoAddr), displayOrder: 4, displayFormat: "X4")]
         public ushort ColorNo {
-            get => (ushort) Data.GetUInt16(_colorNoAddr);
+            get => Data.GetUInt16(_colorNoAddr);
             set => Data.SetUInt16(_colorNoAddr, value);
         }
 
@@ -154,14 +154,14 @@ namespace SF3.Models.Structs.MPD.Model {
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_gouraudShadingTableAddr), displayOrder: 5, displayFormat: "X4")]
         public ushort GouraudShadingTable {
-            get => (ushort) Data.GetUInt16(_gouraudShadingTableAddr);
+            get => Data.GetUInt16(_gouraudShadingTableAddr);
             set => Data.SetUInt16(_gouraudShadingTableAddr, value);
         }
 
         [BulkCopy]
         [TableViewModelColumn(addressField: nameof(_dirAddr), displayOrder: 6, displayFormat: "X4")]
         public ushort Dir {
-            get => (ushort) Data.GetUInt16(_dirAddr);
+            get => Data.GetUInt16(_dirAddr);
             set => Data.SetUInt16(_dirAddr, value);
         }
